chore: move ScanTarget and add sanitization
