+++ internal/analysis/analysis_test.go
@@ -37,6 +37,7 @@ import (
 	httpmocks "github.com/snyk/code-client-go/http/mocks"
 	"github.com/snyk/code-client-go/internal/analysis"
 	"github.com/snyk/code-client-go/observability/mocks"
+	"github.com/snyk/code-client-go/scan"
 )
 
 func setup(t *testing.T) (*confMocks.MockConfig, *httpmocks.MockHTTPClient, *mocks.MockInstrumentor, *mocks.MockErrorReporter, zerolog.Logger) {
@@ -80,7 +81,7 @@ func TestAnalysis_CreateWorkspace(t *testing.T) {
 		Body: io.NopCloser(bytes.NewReader([]byte(`{"data":{"id": "c172d1db-b465-4764-99e1-ecedad03b06a"}}`))),
 	}, nil).Times(1)
 
-	target, err := analysis.NewRepositoryTargetFromPath("../../")
+	target, err := scan.NewRepositoryTargetFromPath("../../")
 	assert.NoError(t, err)
 
 	analysisOrchestrator := analysis.NewAnalysisOrchestrator(mockConfig, &logger, mockHTTPClient, mockInstrumentor, mockErrorReporter)
@@ -98,7 +99,7 @@ func TestAnalysis_CreateWorkspace_NotARepository(t *testing.T) {
 	mockErrorReporter.EXPECT().CaptureError(gomock.Any(), gomock.Any())
 
 	repoDir := t.TempDir()
-	target, err := analysis.NewRepositoryTargetFromPath(repoDir)
+	target, err := scan.NewRepositoryTargetFromPath(repoDir)
 	assert.ErrorContains(t, err, "open local repository")
 
 	analysisOrchestrator := analysis.NewAnalysisOrchestrator(mockConfig, &logger, mockHTTPClient, mockInstrumentor, mockErrorReporter)
@@ -132,7 +133,7 @@ func TestAnalysis_CreateWorkspace_Failure(t *testing.T) {
 		Body: io.NopCloser(bytes.NewReader([]byte(`{"errors": [{"detail": "error detail", "status": "400"}], "jsonapi": {"version": "version"}}`))),
 	}, nil).Times(1)
 
-	target, err := analysis.NewRepositoryTargetFromPath("../../")
+	target, err := scan.NewRepositoryTargetFromPath("../../")
 	assert.NoError(t, err)
 
 	analysisOrchestrator := analysis.NewAnalysisOrchestrator(mockConfig, &logger, mockHTTPClient, mockInstrumentor, mockErrorReporter)
@@ -205,7 +206,7 @@ func TestAnalysis_CreateWorkspace_KnownErrors(t *testing.T) {
 
 			logger := zerolog.Nop()
 
-			target, err := analysis.NewRepositoryTargetFromPath("../../")
+			target, err := scan.NewRepositoryTargetFromPath("../../")
 			assert.NoError(t, err)
 
 			analysisOrchestrator := analysis.NewAnalysisOrchestrator(mockConfig, &logger, mockHTTPClient, mockInstrumentor, mockErrorReporter)
