+++ internal/analysis/mocks/analysis.go
@@ -9,8 +9,8 @@ import (
 	reflect "reflect"
 
 	gomock "github.com/golang/mock/gomock"
-	analysis "github.com/snyk/code-client-go/internal/analysis"
 	sarif "github.com/snyk/code-client-go/sarif"
+	scan "github.com/snyk/code-client-go/scan"
 )
 
 // MockAnalysisOrchestrator is a mock of AnalysisOrchestrator interface.
@@ -37,7 +37,7 @@ func (m *MockAnalysisOrchestrator) EXPECT() *MockAnalysisOrchestratorMockRecorde
 }
 
 // CreateWorkspace mocks base method.
-func (m *MockAnalysisOrchestrator) CreateWorkspace(ctx context.Context, orgId, requestId string, path analysis.ScanTarget, bundleHash string) (string, error) {
+func (m *MockAnalysisOrchestrator) CreateWorkspace(ctx context.Context, orgId, requestId string, path scan.ScanTarget, bundleHash string) (string, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "CreateWorkspace", ctx, orgId, requestId, path, bundleHash)
 	ret0, _ := ret[0].(string)
@@ -64,41 +64,4 @@ func (m *MockAnalysisOrchestrator) RunAnalysis(ctx context.Context, orgId, works
 func (mr *MockAnalysisOrchestratorMockRecorder) RunAnalysis(ctx, orgId, workspaceId interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RunAnalysis", reflect.TypeOf((*MockAnalysisOrchestrator)(nil).RunAnalysis), ctx, orgId, workspaceId)
-}
-
-// MockScannerTarget is a mock of ScanTarget interface.
-type MockScannerTarget struct {
-	ctrl     *gomock.Controller
-	recorder *MockScannerTargetMockRecorder
-}
-
-// MockScannerTargetMockRecorder is the mock recorder for MockScannerTarget.
-type MockScannerTargetMockRecorder struct {
-	mock *MockScannerTarget
-}
-
-// NewMockScannerTarget creates a new mock instance.
-func NewMockScannerTarget(ctrl *gomock.Controller) *MockScannerTarget {
-	mock := &MockScannerTarget{ctrl: ctrl}
-	mock.recorder = &MockScannerTargetMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockScannerTarget) EXPECT() *MockScannerTargetMockRecorder {
-	return m.recorder
-}
-
-// GetPath mocks base method.
-func (m *MockScannerTarget) GetPath() string {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetPath")
-	ret0, _ := ret[0].(string)
-	return ret0
-}
-
-// GetPath indicates an expected call of GetPath.
-func (mr *MockScannerTargetMockRecorder) GetPath() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetPath", reflect.TypeOf((*MockScannerTarget)(nil).GetPath))
 }
